update profile, fix tag permissions, make cw optional, fix message deletion log, add missing language keys, update exp system, add autorole

+++ src/events/cwResponse.js
@@ -10,8 +10,10 @@ module.exports = class extends Event {
 	}
 
 	async run(data) {
+		const guild = this.client.guilds.get(data.message.guild);
+		if (!guild || !guild.settings.get('mod.anti.spam')) return;
 		if (data.scores.overall > 85)
-			this.client.channels
+			guild.channels
 				.get(data.message.channel).messages
 				.get(data.message.id)
 				.delete()
